+++ packages/project/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @kalphq/project
 
+## 0.0.0-dev-20260423070137
+
+### Patch Changes
+
+- dev release
+
+## 0.0.0-dev-20260423065358
+
+### Patch Changes
+
+- dev release
+
+## 0.0.0-dev-20260423060035
+
+### Patch Changes
+
+- dev release
+
+## 0.0.0-dev-20260423051216
+
+### Patch Changes
+
+- dev release
+
 ## 0.0.0-dev-20260423043828
 
 ### Patch Changes